ingestion_trigger_api/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
pydantic
requests
python-dotenv
